Replaced non functional components with functional components
