--- src/classes/LoggedInUser.js
@@ -1,10 +1,0 @@
-let LoggedInUser = {
-  isLoggedIn: false,
-  username: null,
-  email: null,
-  location: null,
-  token: null,
-  sub: null
-};
-
-export default LoggedInUser;
